# backend/src/pilot_space/api/v1/routers/notes.py
"""Notes API router.

CRUD endpoints for notes with AI features integration.
Includes annotation management and discussion threads.

T095: Notes router implementation.
"""

from __future__ import annotations

import logging
from typing import Annotated
from uuid import UUID

from fastapi import APIRouter, HTTPException, Path, Query, status

from pilot_space.api.v1.schemas.annotation import (
    AnnotationListResponse,
    AnnotationResponse,
    AnnotationStatus,
    AnnotationStatusUpdate,
    AnnotationSummary,
)
from pilot_space.api.v1.schemas.base import (
    DeleteResponse,
)
from pilot_space.api.v1.schemas.discussion import (
    DiscussionCreate,
    DiscussionDetailResponse,
    DiscussionListResponse,
    DiscussionResponse,
)
from pilot_space.api.v1.schemas.note import (
    NoteCreate,
    NoteDetailResponse,
    NoteListResponse,
    NotePinUpdate,
    NoteResponse,
    NoteUpdate,
)
from pilot_space.dependencies import CurrentUserId, DbSession

logger = logging.getLogger(__name__)

router = APIRouter(prefix="/notes", tags=["Notes"])


# Note CRUD Endpoints


@router.post(
    "",
    response_model=NoteResponse,
    status_code=status.HTTP_201_CREATED,
    summary="Create a new note",
)
async def create_note(
    note_data: NoteCreate,
    current_user_id: CurrentUserId,
    session: DbSession,
) -> NoteResponse:
    """Create a new note in a project.

    Args:
        note_data: Note creation data.
        current_user_id: Current user ID.
        session: Database session.

    Returns:
        Created note.
    """
    # TODO: Implement note creation service
    # This is a placeholder - actual implementation requires Note model and repository
    raise HTTPException(
        status_code=status.HTTP_501_NOT_IMPLEMENTED,
        detail="Note creation not yet implemented",
    )


@router.get(
    "",
    response_model=NoteListResponse,
    summary="List notes",
)
async def list_notes(
    current_user_id: CurrentUserId,
    session: DbSession,
    project_id: Annotated[UUID | None, Query(description="Filter by project")] = None,
    is_pinned: Annotated[bool | None, Query(description="Filter by pin status")] = None,
    search: Annotated[str | None, Query(description="Search query")] = None,
    cursor: Annotated[str | None, Query(description="Pagination cursor")] = None,
    page_size: Annotated[int, Query(ge=1, le=100, description="Page size")] = 20,
) -> NoteListResponse:
    """List notes with optional filtering and pagination.

    Args:
        current_user_id: Current user ID.
        session: Database session.
        project_id: Optional project filter.
        is_pinned: Optional pin status filter.
        search: Optional search query.
        cursor: Pagination cursor.
        page_size: Items per page.

    Returns:
        Paginated list of notes.
    """
    # TODO: Implement note listing
    raise HTTPException(
        status_code=status.HTTP_501_NOT_IMPLEMENTED,
        detail="Note listing not yet implemented",
    )


@router.get(
    "/{note_id}",
    response_model=NoteDetailResponse,
    summary="Get note details",
)
async def get_note(
    note_id: Annotated[UUID, Path(description="Note ID")],
    current_user_id: CurrentUserId,
    session: DbSession,
) -> NoteDetailResponse:
    """Get detailed note information including content.

    Args:
        note_id: Note ID.
        current_user_id: Current user ID.
        session: Database session.

    Returns:
        Note details with content.
    """
    # TODO: Implement note retrieval
    raise HTTPException(
        status_code=status.HTTP_501_NOT_IMPLEMENTED,
        detail="Note retrieval not yet implemented",
    )


@router.patch(
    "/{note_id}",
    response_model=NoteResponse,
    summary="Update a note",
)
async def update_note(
    note_id: Annotated[UUID, Path(description="Note ID")],
    note_data: NoteUpdate,
    current_user_id: CurrentUserId,
    session: DbSession,
) -> NoteResponse:
    """Update note title or content.

    Args:
        note_id: Note ID.
        note_data: Update data.
        current_user_id: Current user ID.
        session: Database session.

    Returns:
        Updated note.
    """
    # TODO: Implement note update
    raise HTTPException(
        status_code=status.HTTP_501_NOT_IMPLEMENTED,
        detail="Note update not yet implemented",
    )


@router.delete(
    "/{note_id}",
    response_model=DeleteResponse,
    summary="Delete a note",
)
async def delete_note(
    note_id: Annotated[UUID, Path(description="Note ID")],
    current_user_id: CurrentUserId,
    session: DbSession,
) -> DeleteResponse:
    """Soft delete a note.

    Args:
        note_id: Note ID.
        current_user_id: Current user ID.
        session: Database session.

    Returns:
        Delete confirmation.
    """
    # TODO: Implement note deletion
    raise HTTPException(
        status_code=status.HTTP_501_NOT_IMPLEMENTED,
        detail="Note deletion not yet implemented",
    )


@router.patch(
    "/{note_id}/pin",
    response_model=NoteResponse,
    summary="Pin/unpin a note",
)
async def pin_note(
    note_id: Annotated[UUID, Path(description="Note ID")],
    pin_data: NotePinUpdate,
    current_user_id: CurrentUserId,
    session: DbSession,
) -> NoteResponse:
    """Pin or unpin a note.

    Args:
        note_id: Note ID.
        pin_data: Pin status.
        current_user_id: Current user ID.
        session: Database session.

    Returns:
        Updated note.
    """
    # TODO: Implement note pinning
    raise HTTPException(
        status_code=status.HTTP_501_NOT_IMPLEMENTED,
        detail="Note pinning not yet implemented",
    )


# Annotation Endpoints


@router.get(
    "/{note_id}/annotations",
    response_model=AnnotationListResponse,
    summary="Get note annotations",
)
async def get_note_annotations(
    note_id: Annotated[UUID, Path(description="Note ID")],
    current_user_id: CurrentUserId,
    session: DbSession,
    status_filter: Annotated[
        AnnotationStatus | None,
        Query(alias="status", description="Filter by status"),
    ] = None,
    cursor: Annotated[str | None, Query(description="Pagination cursor")] = None,
    page_size: Annotated[int, Query(ge=1, le=100, description="Page size")] = 50,
) -> AnnotationListResponse:
    """Get annotations for a note.

    Args:
        note_id: Note ID.
        current_user_id: Current user ID.
        session: Database session.
        status_filter: Optional status filter.
        cursor: Pagination cursor.
        page_size: Items per page.

    Returns:
        Paginated annotations.
    """
    # TODO: Implement annotation listing
    raise HTTPException(
        status_code=status.HTTP_501_NOT_IMPLEMENTED,
        detail="Annotation listing not yet implemented",
    )


@router.get(
    "/{note_id}/annotations/summary",
    response_model=AnnotationSummary,
    summary="Get annotation summary",
)
async def get_annotation_summary(
    note_id: Annotated[UUID, Path(description="Note ID")],
    current_user_id: CurrentUserId,
    session: DbSession,
) -> AnnotationSummary:
    """Get summary of annotations for a note.

    Args:
        note_id: Note ID.
        current_user_id: Current user ID.
        session: Database session.

    Returns:
        Annotation summary with counts.
    """
    # TODO: Implement annotation summary
    raise HTTPException(
        status_code=status.HTTP_501_NOT_IMPLEMENTED,
        detail="Annotation summary not yet implemented",
    )


@router.patch(
    "/{note_id}/annotations/{annotation_id}",
    response_model=AnnotationResponse,
    summary="Update annotation status",
)
async def update_annotation_status(
    note_id: Annotated[UUID, Path(description="Note ID")],
    annotation_id: Annotated[UUID, Path(description="Annotation ID")],
    status_data: AnnotationStatusUpdate,
    current_user_id: CurrentUserId,
    session: DbSession,
) -> AnnotationResponse:
    """Update an annotation's status (accept, dismiss, convert).

    Args:
        note_id: Note ID.
        annotation_id: Annotation ID.
        status_data: New status.
        current_user_id: Current user ID.
        session: Database session.

    Returns:
        Updated annotation.
    """
    # TODO: Implement annotation status update
    raise HTTPException(
        status_code=status.HTTP_501_NOT_IMPLEMENTED,
        detail="Annotation update not yet implemented",
    )


# Discussion Endpoints


@router.post(
    "/{note_id}/discussions",
    response_model=DiscussionResponse,
    status_code=status.HTTP_201_CREATED,
    summary="Create a discussion",
)
async def create_discussion(
    note_id: Annotated[UUID, Path(description="Note ID")],
    discussion_data: DiscussionCreate,
    current_user_id: CurrentUserId,
    session: DbSession,
) -> DiscussionResponse:
    """Create a new discussion thread on a note.

    Args:
        note_id: Note ID.
        discussion_data: Discussion creation data.
        current_user_id: Current user ID.
        session: Database session.

    Returns:
        Created discussion.
    """
    # Validate note_id matches request
    if discussion_data.note_id != note_id:
        raise HTTPException(
            status_code=status.HTTP_400_BAD_REQUEST,
            detail="note_id in path and body must match",
        )

    # TODO: Implement discussion creation
    raise HTTPException(
        status_code=status.HTTP_501_NOT_IMPLEMENTED,
        detail="Discussion creation not yet implemented",
    )


@router.get(
    "/{note_id}/discussions",
    response_model=DiscussionListResponse,
    summary="List note discussions",
)
async def list_discussions(
    note_id: Annotated[UUID, Path(description="Note ID")],
    current_user_id: CurrentUserId,
    session: DbSession,
    cursor: Annotated[str | None, Query(description="Pagination cursor")] = None,
    page_size: Annotated[int, Query(ge=1, le=50, description="Page size")] = 20,
) -> DiscussionListResponse:
    """List discussions for a note.

    Args:
        note_id: Note ID.
        current_user_id: Current user ID.
        session: Database session.
        cursor: Pagination cursor.
        page_size: Items per page.

    Returns:
        Paginated discussions.
    """
    # TODO: Implement discussion listing
    raise HTTPException(
        status_code=status.HTTP_501_NOT_IMPLEMENTED,
        detail="Discussion listing not yet implemented",
    )


@router.get(
    "/{note_id}/discussions/{discussion_id}",
    response_model=DiscussionDetailResponse,
    summary="Get discussion details",
)
async def get_discussion(
    note_id: Annotated[UUID, Path(description="Note ID")],
    discussion_id: Annotated[UUID, Path(description="Discussion ID")],
    current_user_id: CurrentUserId,
    session: DbSession,
) -> DiscussionDetailResponse:
    """Get discussion with all comments.

    Args:
        note_id: Note ID.
        discussion_id: Discussion ID.
        current_user_id: Current user ID.
        session: Database session.

    Returns:
        Discussion with comments.
    """
    # TODO: Implement discussion retrieval
    raise HTTPException(
        status_code=status.HTTP_501_NOT_IMPLEMENTED,
        detail="Discussion retrieval not yet implemented",
    )


__all__ = ["router"]