# specs/001-pilot-space-mvp/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Pilot Space API
  description: |
    AI-augmented SDLC platform API for project management with intelligent features.

    ## Authentication
    All endpoints require Bearer token authentication via Supabase Auth (GoTrue).
    Authorization is enforced via Row-Level Security (RLS) policies.

    ## Rate Limits
    - Standard endpoints: 1000 requests/minute per workspace
    - AI endpoints: 100 requests/minute per workspace (subject to LLM provider limits)
    - Webhook endpoints: 500 requests/minute

    ## Pagination
    List endpoints support cursor-based pagination using `cursor` and `limit` parameters.
    Response format: `{ data: [...], meta: { total, cursor, hasMore } }`

    ## Updated
    Session 2026-01-22: Migrated to Supabase platform
  version: 2.0.0
  contact:
    name: Pilot Space Team
    url: https://pilot-space.com
  license:
    name: MIT
servers:
  - url: https://api.pilot-space.com/api/v1
    description: Production
  - url: http://localhost:8000/api/v1
    description: Development
...
paths:
...
  /workspaces/{workspaceSlug}/notes:
    parameters:
      - $ref: "#/components/parameters/workspaceSlug"
    get:
      tags: [Notes]
      summary: List notes in workspace
      operationId: listNotes
      security:
        - bearerAuth: []
      parameters:
        - name: projectId
          in: query
          schema:
            type: string
            format: uuid
        - name: isPinned
          in: query
          schema:
            type: boolean
        - $ref: "#/components/parameters/cursor"
        - $ref: "#/components/parameters/limit"
      responses:
        "200":
          description: List of notes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedNotes"

    post:
      tags: [Notes]
      summary: Create note
      operationId: createNote
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNoteRequest"
      responses:
        "201":
          description: Created note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
  /workspaces/{workspaceSlug}/notes/{noteId}:
    parameters:
      - $ref: "#/components/parameters/workspaceSlug"
      - $ref: "#/components/parameters/noteId"
    get:
      tags: [Notes]
      summary: Get note with content
      operationId: getNote
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Note with full content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NoteDetail"

    patch:
      tags: [Notes]
      summary: Update note
      operationId: updateNote
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateNoteRequest"
      responses:
        "200":
          description: Updated note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NoteDetail"

    delete:
      tags: [Notes]
      summary: Delete note (soft delete)
      operationId: deleteNote
      security:
        - bearerAuth: []
      responses:
        "204":
          description: Note deleted
...
  /workspaces/{workspaceSlug}/notes/{noteId}/annotations:
    parameters:
      - $ref: "#/components/parameters/workspaceSlug"
      - $ref: "#/components/parameters/noteId"
    get:
      tags: [Notes]
      summary: Get note annotations
      operationId: getNoteAnnotations
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of annotations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NoteAnnotation"
...
  /workspaces/{workspaceSlug}/notes/{noteId}/discussions:
    parameters:
      - $ref: "#/components/parameters/workspaceSlug"
      - $ref: "#/components/parameters/noteId"
    get:
      tags: [Notes]
      summary: Get threaded discussions
      operationId: getNoteDiscussions
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of discussions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ThreadedDiscussion"

    post:
      tags: [Notes]
      summary: Start new discussion thread
      operationId: createDiscussion
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDiscussionRequest"
      responses:
        "201":
          description: Discussion created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ThreadedDiscussion"

  # ============================================================================
  # Issues
  # ============================================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    workspaceSlug:
      name: workspaceSlug
      in: path
      required: true
      schema:
        type: string
        pattern: "^[a-z0-9-]+$"
...
    noteId:
      name: noteId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    cursor:
      name: cursor
      in: query
      schema:
        type: string
    limit:
      name: limit
      in: query
      schema:
        type: integer
        default: 50
        maximum: 100
  schemas:
...
    StateGroup:
      type: string
      enum: [backlog, unstarted, started, completed, cancelled]

    # ==================== User ====================
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        displayName:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        avatarUrl:
          type: string
          format: uri
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
...
    Note:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        emoji:
          type: string
        projectId:
          type: string
          format: uuid
        isPinned:
          type: boolean
        syncState:
          type: string
          enum: [draft, synced, conflict]
        createdBy:
          $ref: "#/components/schemas/User"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    NoteDetail:
      allOf:
        - $ref: "#/components/schemas/Note"
        - type: object
          properties:
            content:
              type: object
              description: TipTap/ProseMirror JSON content
            annotations:
              type: array
              items:
                $ref: "#/components/schemas/NoteAnnotation"
            linkedIssues:
              type: array
              items:
                $ref: "#/components/schemas/IssueLink"
    CreateNoteRequest:
      type: object
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 255
        emoji:
          type: string
        projectId:
          type: string
          format: uuid
        templateId:
          type: string
          format: uuid
        content:
          type: object
    UpdateNoteRequest:
      type: object
      properties:
        title:
          type: string
        emoji:
          type: string
        content:
          type: object
        isPinned:
          type: boolean
    NoteAnnotation:
      type: object
      properties:
        id:
          type: string
          format: uuid
        blockId:
          type: string
        annotationType:
          type: string
          enum: [suggestion, issue_candidate, warning, linked_issue, comment]
        content:
          type: string
        confidence:
          type: number
          minimum: 0
          maximum: 1
        metadata:
          type: object
        isDismissed:
          type: boolean
        isAccepted:
          type: boolean
        createdAt:
          type: string
          format: date-time
    PaginatedNotes:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Note"
        cursor:
          type: string
        hasMore:
          type: boolean
...
    ThreadedDiscussion:
      type: object
      properties:
        id:
          type: string
          format: uuid
        blockId:
          type: string
        isResolved:
          type: boolean
        comments:
          type: array
          items:
            $ref: "#/components/schemas/DiscussionComment"
        createdAt:
          type: string
          format: date-time
    DiscussionComment:
      type: object
      properties:
        id:
          type: string
          format: uuid
        content:
          type: string
        author:
          $ref: "#/components/schemas/User"
        isEdited:
          type: boolean
        createdAt:
          type: string
          format: date-time
    CreateDiscussionRequest:
      type: object
      required: [blockId, content]
      properties:
        blockId:
          type: string
        content:
          type: string

    # ==================== Issue ====================
...
    IssueLink:
      type: object
      properties:
        issueId:
          type: string
          format: uuid
        identifier:
          type: string
        name:
          type: string
        state:
          $ref: "#/components/schemas/State"
...
    State:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        color:
          type: string
        group:
          $ref: "#/components/schemas/StateGroup"
...